Added context menu for each class on long press
Now we can delete classes from the list via the context menu, and soon
edit. Also, I using a new fragment MessageFragment for displaying
messages that replaces QuestionFragment.. its basically the same, but It
has a static newInstance method that came with the blank fragment
template I used that looks a little cleaner.