prepare for "final" product -- holo-ized things
liked the grid idea, so made it more Holo / cards themed

also regressed back to the drawer menu and made it more of a focal
point, but could probably afford another set of eyes on how intuitive it
turned out. putting them on the main page seems to makes a lot of sense, as
well, since our home page doesn't do much else. My main reason was that
students might take 2 quizzes at once (yeah, edge case and we won't ever scale sooo)

I think the only thing I missed was deletion of a class. Obviously this
is a pretty nice thing to have, granted not vital for presentation but
for actual usability it is. If we can think of a nice way to get that in
quickly then we should probably try to do that. in the drawer I think
would be the best place, or a separate settings page with a list of
classes to edit / delete... actually, think I removed InfoEntryFragment
(if you do a git checkout of file from HEAD~2 you should get it back). I
removed all files that weren't getting used so that got looped in
(woops). uh, should probably get that back in some form just need to
integrate it.