Got delete back in, renaming class kinda
OK so I got the ability to delete a class working again via a context
menu on a long press. That menu also lets you view the class info
(title, cId, sId) and edit the title, however renaming is buggy. It
correctly renames the class but I couldn't get the drawer to update and
show new name. It only shows the new name after a app restart. The bug
is that if you rename a class and then try to delete that class before
an app restart it will crash. So as far as the presentation we should
just not rename any classes unless you can get the drawer to update
correctly.

Also, there may be merge conflicts... but the only files I changed were
MainActivity (added context menu stuff) and fragment_question (had to
change an ID tag to build). Also readded ClassInfoDialogFragment since
it is needed to show class info from context menu.